+++ src/components/DataTable/plugins/Subrow/subrowReducer.ts
@@ -1,4 +1,4 @@
-export const SUBROW_TOGGLE = "SUBROW_TOGGLE";
+export const SUBROW_TOGGLE = 'SUBROW_TOGGLE';
 
 interface SubrowState {
     open: Array<number>;
@@ -9,7 +9,6 @@ export const initialState: SubrowState = {
 };
 
 const reducer = (state = initialState, action: any): SubrowState => {
-
     let tmp;
 
     switch (action.type) {
@@ -18,7 +17,7 @@ const reducer = (state = initialState, action: any): SubrowState => {
             if (tmp.indexOf(action.payload) !== -1) {
                 tmp.splice(tmp.indexOf(action.payload), 1);
             } else {
-                tmp.push (action.payload)
+                tmp.push(action.payload);
             }
 
             return {
@@ -30,4 +29,4 @@ const reducer = (state = initialState, action: any): SubrowState => {
     }
 };
 
-export const subrowReducer = {reducer, initialState}
+export const subrowReducer = { reducer, initialState };