+++ src/components/Gantt/internal/Tooltip.tsx
@@ -115,7 +115,7 @@ export const StandardTooltipContent: React.FC<{
     };
     return (
         <TooltipDefaultContainer style={style}>
-            <b style={{ fontSize: fontSize + 6 }}>{`${task.name}: ${task.start.getDate()}-${
+            <b>{`${task.name}: ${task.start.getDate()}-${
                 task.start.getMonth() + 1
             }-${task.start.getFullYear()} - ${task.end.getDate()}-${
                 task.end.getMonth() + 1