+++ src/components/Gantt/types/gantt-task-actions.ts
@@ -1,18 +1,18 @@
-import { TaskBar } from "./bar-task";
+import { TaskBar } from './bar-task';
 
-export type BarMoveAction = "progress" | "end" | "start" | "move";
+export type BarMoveAction = 'progress' | 'end' | 'start' | 'move';
 export type GanttContentMoveAction =
-  | "mouseenter"
-  | "mouseleave"
-  | "delete"
-  | "dblclick"
-  | "click"
-  | "select"
-  | ""
-  | BarMoveAction;
+    | 'mouseenter'
+    | 'mouseleave'
+    | 'delete'
+    | 'dblclick'
+    | 'click'
+    | 'select'
+    | ''
+    | BarMoveAction;
 
 export type GanttEvent = {
-  changedTask?: TaskBar;
-  originalSelectedTask?: TaskBar;
-  action: GanttContentMoveAction;
+    changedTask?: TaskBar;
+    originalSelectedTask?: TaskBar;
+    action: GanttContentMoveAction;
 };